Fixes #726 Mockito Public API Support Fix tests that are failed only on travis due different order of executing.
