+++ build.gradle
@@ -57,7 +57,6 @@ allprojects{
 }
 
 apply from: "${gradleScriptDir}/modules.gradle"
-//apply from: "${gradleScriptDir}/release/release.gradle"
 
 jar{
     enabled = false
