+++ powermock-api/powermock-api-mockito2/src/main/java/org/powermock/api/mockito/mockmaker/PowerMockMaker.java
@@ -37,7 +37,7 @@ public class PowerMockMaker implements MockMaker {
     }
     
     @Override
-    public <T> T createMock(MockCreationSettings<T> settings, MockHandler handler) {
+    public <T> T createMock(MockCreationSettings<T> settings, MockHandler handler){
         return mockMaker.createMock(settings, handler);
     }
