+++ powermock-api/powermock-api-mockito2/src/test/java/org/powermock/api/mockito/internal/expectation/DefaultMethodExpectationSetupTestCase.java
@@ -23,7 +23,7 @@ import org.powermock.reflect.Whitebox;
 /**
  * @author Stanislav Chizhov
  */
-public class DefaultMethodExpectationSetupTest {
+public class DefaultMethodExpectationSetupTestCase {
 
     private final CUT object = new CUT();
